backend/requirements-tools.txt
# Modern Document Conversion Tools (2025)
# Core FastAPI (already in main requirements.txt)

# Document Conversion - Modern Stack
pypdf==4.0.1              # Latest PDF library (replaces PyPDF2)
python-docx==1.1.0        # Word document handling
docx2pdf==0.1.8           # Word to PDF conversion (Windows)
pdf2docx==0.5.8           # PDF to Word conversion
python-pptx==0.6.23       # PowerPoint handling
openpyxl==3.1.2           # Excel handling (modern)
unoserver==3.5            # LibreOffice API server (universal converter) - Latest stable

# PDF Tools - Advanced
pypdfium2==4.26.0         # Fast PDF rendering (Google's PDFium)
pikepdf==8.11.0           # PDF manipulation (better than PyPDF2)
reportlab==4.0.9          # PDF generation from scratch

# Image Processing - Modern AI
Pillow==10.2.0            # Latest Pillow
rembg==2.0.56             # AI background removal (SOTA)
opencv-python==4.9.0.80   # Latest OpenCV
pillow-heif==0.15.0       # HEIF/HEIC support (iPhone photos)

# OCR - Deep Learning
easyocr==1.7.1            # Deep learning OCR (Vietnamese + 80 languages)
paddleocr==2.7.3          # Baidu's SOTA OCR (alternative)

# Text Processing - Modern NLP
spacy==3.7.2              # Modern NLP (faster than NLTK)
transformers==4.37.2      # Hugging Face transformers (AI models)

# File Processing
python-magic-bin==0.4.14  # File type detection (Windows binary included)
filetype==1.2.0           # Fast file type detection

# Async & Performance
aiofiles==23.2.1          # Async file I/O
httpx==0.26.0             # Async HTTP client

# Monitoring
psutil==5.9.8             # System monitoring
